Add Security

+++ Market/Controllers/AnuncioController.cs
@@ -5,6 +5,7 @@ using System;
 using System.Collections.Generic;
 using System.Linq;
 using System.Threading.Tasks;
+using WebApp.Market.Middleware;
 
 namespace WebApp.Market.Controllers
 {
@@ -12,18 +13,29 @@ namespace WebApp.Market.Controllers
     {
         private readonly IAnuncioService _anuncioService;
         private readonly ICategoriaService _categoriaService;
-        public AnuncioController(IAnuncioService anuncioService, ICategoriaService  categoriaService)
+        private readonly UserSession _userSession;
+        public AnuncioController(IAnuncioService anuncioService, ICategoriaService  categoriaService,UserSession userSession)
         {
             _anuncioService = anuncioService;
             _categoriaService = categoriaService;
+            _userSession = userSession;
         }
         public async Task<IActionResult> Index()
         {
-            var list = await _anuncioService.GetAllViewModel();
-            return View(list);
+            if (!_userSession.hasUser())
+            {
+                var list = await _anuncioService.GetAllUserViewModel();
+                return View(list);
+            }
+            return RedirectToRoute(new { controller = "Usuario", action = "Index" });
         }
         public async Task<IActionResult> Create()
         {
+            if (!_userSession.hasUser())
+            {
+                return RedirectToRoute(new { controller = "Usuario", action = "Index" });
+            }
+
             SaveAnuncioViewModel anuncioViewModel = new SaveAnuncioViewModel();
             anuncioViewModel.Categorias = await _categoriaService.GetAllViewModel();
             return View("SaveAnuncio", anuncioViewModel);
@@ -31,6 +43,11 @@ namespace WebApp.Market.Controllers
         [HttpPost]
         public async Task<IActionResult> Create(SaveAnuncioViewModel anuncioViewModel)
         {
+            if (!_userSession.hasUser())
+            {
+                return RedirectToRoute(new { controller = "Usuario", action = "Index" });
+            }
+
             if (!ModelState.IsValid)
             {
                 anuncioViewModel.Categorias = await _categoriaService.GetAllViewModel();
@@ -42,6 +59,11 @@ namespace WebApp.Market.Controllers
         }
         public async Task<IActionResult> Edit(int id)
         {
+            if (!_userSession.hasUser())
+            {
+                return RedirectToRoute(new { controller = "Usuario", action = "Index" });
+            }
+
             SaveAnuncioViewModel anuncioViewModel = await _anuncioService.GetByIdSaveViewModel(id);
             anuncioViewModel.Categorias = await _categoriaService.GetAllViewModel();
             return View("SaveAnuncio", anuncioViewModel);
@@ -49,6 +71,11 @@ namespace WebApp.Market.Controllers
         [HttpPost]
         public async Task<IActionResult> Edit(SaveAnuncioViewModel saveAnuncioViewModel)
         {
+            if (!_userSession.hasUser())
+            {
+                return RedirectToRoute(new { controller = "Usuario", action = "Index" });
+            }
+
             if (!ModelState.IsValid)
             {
                 saveAnuncioViewModel.Categorias = await _categoriaService.GetAllViewModel();
@@ -60,17 +87,32 @@ namespace WebApp.Market.Controllers
         }
         public async Task<IActionResult> Delete(int id)
         {
+            if (!_userSession.hasUser())
+            {
+                return RedirectToRoute(new { controller = "Usuario", action = "Index" });
+            }
+
             return View(await _anuncioService.GetByIdSaveViewModel(id));
         }
         [HttpPost]
         public async Task<IActionResult> DeletePost(int id)
         {
+            if (!_userSession.hasUser())
+            {
+                return RedirectToRoute(new { controller = "Usuario", action = "Index" });
+            }
+
             await _anuncioService.Delete(id);
             return RedirectToRoute(new { controller = "Anuncio", action ="Index" });
 
         }
         public async Task<IActionResult> Detail(int id)
         {
+            if (!_userSession.hasUser())
+            {
+                return RedirectToRoute(new { controller = "Usuario", action = "Index" });
+            }
+
             return View(await _anuncioService.GetByIdSaveViewModel(id));
         }
     }
